start: new version off go client
