+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/kenriortega/qvapay-go
 
-go 1.17
+go 1.18
 
 require github.com/stretchr/testify v1.7.0
