+++ merchants_test.go
@@ -19,7 +19,7 @@ const (
 
 func Test_Invalid_URL(t *testing.T) {
 
-	api := qvapay.NewAPIClient(appID, secretID, "ht&@-tp://:aa", true, nil, nil)
+	api := qvapay.NewPaymentAppClient(appID, secretID, "ht&@-tp://:aa", true, nil, nil)
 
 	actual, err := api.GetInfo(context.Background())
 	assert.Error(t, err)
@@ -50,7 +50,7 @@ func Test_Get_Info(t *testing.T) {
 	)
 	defer s.Close()
 
-	client := qvapay.NewAPIClient(appID, secretID, s.URL, true, nil, nil)
+	client := qvapay.NewPaymentAppClient(appID, secretID, s.URL, true, nil, nil)
 	info, err := client.GetInfo(context.Background())
 	if err != nil {
 		t.Fatalf(err.Error())
@@ -78,7 +78,7 @@ func Test_Create_Invoice(t *testing.T) {
 		}),
 	)
 	defer s.Close()
-	client := qvapay.NewAPIClient(appID, secretID, s.URL, true, nil, nil)
+	client := qvapay.NewPaymentAppClient(appID, secretID, s.URL, true, nil, nil)
 	invoice, err := client.CreateInvoice(context.Background(), amountInput, "Enanitos verdes", "BRID56568989")
 	if err != nil {
 		t.Fatalf(err.Error())
@@ -129,7 +129,7 @@ func Test_Get_Txs(t *testing.T) {
 		}),
 	)
 	defer s.Close()
-	client := qvapay.NewAPIClient(appID, secretID, s.URL, true, nil, nil)
+	client := qvapay.NewPaymentAppClient(appID, secretID, s.URL, true, nil, nil)
 	query := qvapay.APIQueryParams{}
 	txs, err := client.GetTransactions(context.Background(), query)
 	if err != nil {
@@ -189,7 +189,7 @@ func Test_Get_Tx(t *testing.T) {
 		}),
 	)
 	defer s.Close()
-	client := qvapay.NewAPIClient(appID, secretID, s.URL, true, nil, nil)
+	client := qvapay.NewPaymentAppClient(appID, secretID, s.URL, true, nil, nil)
 	tx, err := client.GetTransaction(context.Background(), inputId)
 	if err != nil {
 		t.Fatalf(err.Error())
@@ -211,7 +211,7 @@ func Test_Get_Balance(t *testing.T) {
 	)
 	defer s.Close()
 
-	client := qvapay.NewAPIClient(appID, secretID, s.URL, true, nil, nil)
+	client := qvapay.NewPaymentAppClient(appID, secretID, s.URL, true, nil, nil)
 	balance, err := client.GetBalance(context.Background())
 	if err != nil {
 		t.Fatalf(err.Error())
